MOB stats upgrade manager

+++ Unity Project/Assets/Jesus Assets/AI and Room Transition ScriptsScripts/RangedEnemyScripts/ProjectileScript.cs
@@ -6,7 +6,7 @@ public class Projectile : MonoBehaviour
 {
     public float speed = 10f;             // Speed of the projectile
     public float lifespan = 3f;           // Time in seconds before the projectile is destroyed
-    private int damage = 10;              // Damage inflicted on the player, can be set by enemy
+    private float damage = 10;              // Damage inflicted on the player, can be set by enemy
 
     private Vector2 direction;            // Direction to travel
 
@@ -29,7 +29,7 @@ public class Projectile : MonoBehaviour
     }
 
     // Method to set the damage of the projectile, which can be called by the enemy script
-    public void SetDamage(int newDamage)
+    public void SetDamage(float newDamage)
     {
         damage = newDamage;
     }
